+++ src/components/SignUp/SignUp.css
@@ -1,0 +1,10 @@
+.login-text{
+    color: #00CED1;
+    font-family: 'Fuzzy Bubbles', cursive;
+}
+
+.login-wrapper {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
